Rename SnapshotManager to RsyncSnapshot and move into helpers
Also adding brief description in docs
